this almost works on EE
for some reason on EE and desktop the db becomes active and queries to
create schema starts working but it fails with db manager issues error
after some time. This can be mitigated in both places by just
terminating liberty pod. That ends up creating new liberty pod which
runs the queries again and that works.